Remove creature ambiguity
`Creature` and `CreatureBuilder` belong in the core module.
